Respect `use_interpolation` with `use_polars`
Existing code interpolates signal readings whenever `use_polars` is
True, because Polars `Series` objects do not have an index and are not
automatically aligned in the final DataFrame creation.

This change has `data` collect up two-column `DataFrame` objects, rather
than `Series`, so that the timestamp for each signal is retained. The
frames are then (lazily) joined on the `timestamps` column.

That allows the check for `use_interpolation` to ignore the value of
`use_polars` when deciding whether to interpolate signals.

As part of this change, checks are also removed for `use_polars` in
combination with `interpolation_outwards_with_nan`. This check is no
longer needed as we now support null-filling with polars.